Updates default configuration to include local.php.dist template
Adds a local.php user configuration template called `local.php.dist`. This file is to
include any local credentials specific to the user's credentials and is set to be
ignored by the VCS.

Documentation will need to be added with instructions to add credentials to this
file and then renamed `local.php`.